+++ cloud_vfs/__init__.py
@@ -1,3 +1,3 @@
 """Manual Azure blob virtual filesystem."""
 
-__version__ = "0.5.8"
+__version__ = "0.5.9"